% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biocard.R
\name{get_biocard}
\alias{get_biocard}
\title{Get BIOCARD Data}
\usage{
get_biocard(path)
}
\arguments{
\item{path}{The path where the biocard data stored}
}
\value{
a table with merged biocard data
}
\description{
Get BIOCARD Data
}
\examples{
\dontrun{
path = "/Users/name/Documents/R/adbiomarker/Data/BIOCARD"
dt_biocard <- get_biocard(path)
}
}
